+++ src/main/java/com/example/demo/repository/ElectionCandidateRepository.java
@@ -25,10 +25,10 @@ public interface ElectionCandidateRepository extends JpaRepository<ElectionCandi
     List<ElectionCandidate>  findByElectionId(@Param("electionId") Long electionId);
 
     @Query(value = "SELECT e.candidate_id FROM election_candidates e where e.election_id = ?1 and e.competing_in_locality = ?2 and e.candidate_type_id=?3", nativeQuery = true)
-    List<Long>  findByElectionIdAndCompetingInLocality(@Param("electionId") Long electionId, @Param("competingInLocality") Integer competingInLocality, @Param("candidateTypeId") Integer candidateTypeId);
+    List<Long>  findByElectionIdAndCompetingInLocality(@Param("electionId") Long electionId, @Param("competingInLocality") Long competingInLocality, @Param("candidateTypeId") Long candidateTypeId);
 
     @Query(value = "SELECT e.candidate_id FROM election_candidates e where e.election_id = ?1 and e.candidate_type_id=?2 and e.county = ?3", nativeQuery = true)
-    List<Long>  findByElectionIdAndCompetingInCounty(@Param("electionId") Long electionId, @Param("candidateTypeId") Integer candidateTypeId, @Param("county") String county);
+    List<Long>  findByElectionIdAndCompetingInCounty(@Param("electionId") Long electionId, @Param("candidateTypeId") Long candidateTypeId, @Param("county") String county);
 
     @Query(value = "SELECT e.candidate_id FROM election_candidates e where e.election_id = ?1 and e.candidate_type_id = ?2", nativeQuery = true)
     List<Long>  findByElectionIdAndCandidateTypeId(@Param("electionId") Long electionId, @Param("candidateTypeId") Long candidateTypeId);
