+++ src/main/java/com/example/demo/repository/LocalityRepository.java
@@ -20,7 +20,7 @@ public interface LocalityRepository extends JpaRepository<Locality, Integer> {
     List<Map<String, Object>> findLocalitiesForMap();
 
     @Query(value = "SELECT id, name, county, population, zip, lat, lng, auto, diacritics FROM localities WHERE id = ?1", nativeQuery = true)
-    Optional<Locality> findById(@Param("localityId") Integer localityId);
+    Optional<Locality> findById(@Param("localityId") Long localityId);
 
     Optional<Locality> findByName(String name);
 }
